requirements.txt
scikit-learn
pandas
numpy
langchain
langgraph
langchain-community
langchain-google-genai
pypdf
pillow
python-dotenv